refactor changes boy Scot
