+++ src/ai_hawk/authenticator.py
