+++ src/ai_hawk/bot_facade.py
