+++ src/ai_hawk/job_manager.py
@@ -9,10 +9,12 @@ from inputimeout import inputimeout, TimeoutOccurred
 from selenium.common.exceptions import NoSuchElementException
 from selenium.webdriver.common.by import By
 
+
+from ai_hawk.linkedIn_easy_applier import AIHawkEasyApplier
 import src.utils as utils
 from app_config import MINIMUM_WAIT_TIME
 from src.job import Job
-from src.aihawk_easy_applier import AIHawkEasyApplier
+
 from loguru import logger
 import urllib.parse
