+++ tests/test_aihawk_job_manager.py
@@ -3,7 +3,7 @@ from unittest import mock
 from pathlib import Path
 import os
 import pytest
-from src.aihawk_job_manager import AIHawkJobManager
+from ai_hawk.job_manager import AIHawkJobManager
 from selenium.common.exceptions import NoSuchElementException
 from loguru import logger
